Changing the footer based on the role assigned. (Not working yet.)
